Support stemcell name property
Broadcom has started supporting FIPS enabled stemcells whose version
matches the non-FIPS stemcells and without a name specified the FIPS may
unintentionally be used when a deployment might not support FIPS.

To avoid this allow consumers to set a stemcell name property to point
to the expected stemcell name. This allows BOSH to ensure the desired
stemcell is always used and not unintentionally deploying with the wrong
variant.